Add a demo page ready to deploy to GH pages
- Points integration tests to new /test route
- Adds Bootstrap button styles to dummy app CSS
