+++ tests/dummy/app/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Dummy</title>
+    <title>Ember CLI Flash</title>
     <meta name="description" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
