+++ tests/dummy/app/templates/application.hbs
@@ -1,6 +1,4 @@
-<h1 id="title">Ember-CLI-Flash Demo</h1>
-
-<p>Simple, highly configurable flash messages for ember-cli.</p>
+{{outlet}}
 
 {{#each flashMessages.arrangedQueue as |flash|}}
   {{#flash-message flash=flash as |component flash|}}
