+++ tests/dummy/app/templates/test.hbs
@@ -1,0 +1,3 @@
+<h1 id="title">Ember-CLI-Flash Demo</h1>
+
+<p>Simple, highly configurable flash messages for ember-cli.</p>
