+++ tests/dummy/config/environment.js
@@ -4,7 +4,7 @@ module.exports = function(environment) {
   var ENV = {
     modulePrefix: 'dummy',
     environment: environment,
-    rootURL: '/',
+    rootURL: '/ember-cli-flash/',
     locationType: 'auto',
     EmberENV: {
       FEATURES: {
